requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
peewee==3.15.4
pyTelegramBotAPI==4.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot-calendar==1.0.5
requests==2.26.0
six==1.16.0
urllib3==1.26.7
